main.py
#!/usr/bin/env python3
"""
Main entry point for the insurance taxonomy classifier.
"""
import argparse
import logging
import yaml
import os

def main():
    """Main function of the application."""
    parser = argparse.ArgumentParser(description="Insurance Taxonomy Classifier")
    parser.add_argument("--config", type=str, default="config/config.yaml",
                        help="Path to configuration file")
    parser.add_argument("--mode", type=str, choices=["train", "evaluate", "predict"], 
                        default="train", help="Operating mode")
    parser.add_argument("--input_file", type=str, help="Input file for predictions")
    parser.add_argument("--output_file", type=str, help="Output file for results")
    args = parser.parse_args()
    
    # Configure logging
    logging.basicConfig(
        level=logging.INFO,
        format="%(asctime)s - %(name)s - %(levelname)s - %(message)s"
    )
    logger = logging.getLogger("insurance_classifier")
    
    # Load configuration
    with open(args.config, "r") as f:
        config = yaml.safe_load(f)
    
    logger.info(f"Running in mode: {args.mode}")
    
    if args.mode == "train":
        # Implement training logic
        pass
    elif args.mode == "evaluate":
        # Implement evaluation logic
        pass
    elif args.mode == "predict":
        # Implement prediction logic
        if not args.input_file or not args.output_file:
            logger.error("For predict mode, --input_file and --output_file are required")
            return
        pass
    
    logger.info("Process completed successfully")

if __name__ == "__main__":
    main()
