requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
nltk>=3.6.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.60.0
PyYAML>=6.0
pytest>=6.2.0
jupyter>=1.0.0
torch>=1.9.0
